Fixed casting issue from unsigned long to int32_t

+++ EZAudio/TPCircularBuffer.c
@@ -46,7 +46,7 @@ bool TPCircularBufferInit(TPCircularBuffer *buffer, int length) {
     int retries = 3;
     while ( true ) {
 
-        buffer->length = round_page(length);    // We need whole page sizes
+        buffer->length = (int32_t)round_page(length);    // We need whole page sizes
 
         // Temporarily allocate twice the length, so we have the contiguous address space to
         // support a second instance of the buffer directly after
